requirements.txt
moviepy==1.0.3
gTTS==2.5.0
Pillow==10.2.0
numpy==1.26.3
pydub==0.25.1
requests==2.31.0
edge-tts==6.1.9
